better review add

+++ reviews/models.py
@@ -4,7 +4,7 @@ from rest_framework import serializers
 class Review(models.Model):
     user = models.ForeignKey('users.User', on_delete=models.CASCADE)
     game = models.ForeignKey('games.Game', on_delete=models.CASCADE)
-    review = models.IntegerField(blank=True, null=True)
+    review = models.FloatField(blank=True, null=True)
 
     class Meta:
         unique_together = [['user', 'game']]
